requirements.txt
numpy~=1.24.2
networkx~=2.8.8
matplotlib~=3.7.0
nbformat~=5.7.3
nbconvert~=7.2.9
dimod~=0.12.3
pygad~=2.19.2
dwave-ocean-sdk>=3.3.0
